+++ package/sbp_protocol/src/framer_sbp.c
@@ -34,11 +34,11 @@ typedef struct {
   uint8_t send_buffer[SBP_MSG_LEN_MAX];
 } framer_sbp_state_t;
 
-static u32 sbp_read(u8 *buff, u32 n, void *context)
+static ssize_t sbp_read(u8 *buff, u32 n, void *context)
 {
   sbp_io_context_t *c = (sbp_io_context_t *)context;
 
-  u32 count = c->read_length - c->read_offset;
+  ssize_t count = c->read_length - c->read_offset;
   if (count > n) {
     count = n;
   }
@@ -48,11 +48,11 @@ static u32 sbp_read(u8 *buff, u32 n, void *context)
   return count;
 }
 
-static u32 sbp_write(u8 *buff, u32 n, void *context)
+static ssize_t sbp_write(u8 *buff, u32 n, void *context)
 {
   sbp_io_context_t *c = (sbp_io_context_t *)context;
 
-  u32 count = c->write_length - c->write_offset;
+  ssize_t count = c->write_length - c->write_offset;
   if (count > n) {
     count = n;
   }
